// src/apis/types.ts
export type EnvListItemType = {
  label: string | number
  value: string | number
}

export type EnergyListItemType = {
  label: string | number
  value: string | number
}

export type StatisticsListItemType = {
  label: string | number
  value: string | number
  type: 'danger' | 'normal' | 'warnning'
}

export type YawAngleListItemType = {
  label: string | number
  value: string | number
}

export type ActiveListItemType = {
  label: string | number
  value: string | number
}

export type ErrorListItemType = {
  label: string | number
  value: string | number
  type: 'danger' | 'normal' | 'warnning'
}